requirements.txt
xgboost
dnspython
Flask-BabelEx
flask-mongoengine
Flask-RESTful
gunicorn
python-dotenv
PyYAML
numpy
dnspython
pymongo
Flask
gunicorn
Jinja2
Werkzeug
numpy
pandas
scipy
scikit-learn
matplotlib
kaggle


